SimpLL: Move part of main to a separate function.
There are now two interfaces to SimpLL (library and command line),
therefore there is a part of code common for both. This part was moved
to a separate function named processAndCompare.

In addition Transforms.cpp/h was renamed to ModuleAnalysis.cpp/h to
better describe its contents.
